Add possibility to choose port or just build the site
I moved the main function to its own page, where it loads configuration and accept arguments
to have different behaviors.

Launching it without arguments will result in just building the page, with -server launch the local server
at a default port and with -port you can choose the port.

Changed the name of some files to have a cooler name for the projectwq